pyproject.toml
[tool.poetry]
name = "csv2docx"
version = "0.1.0"
description = "Generates .docx files from .csv files using a .docx template."
authors = ["David Verweij <davidverweij@gmail.com>", "Jay Rainey <hi@jawrainey.me>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/davidverweij/csv2docx"
repository = "https://github.com/davidverweij/csv2docx"

[tool.poetry.dependencies]
python = "^3.8.0"
docx-mailmerge = "^0.5.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
coverage = {version = "^5.1", extras = ["toml"]}
flake8 = "^3.7.9"
flake8-aaa = "^0.9.0"
flake8-bandit = "^2.1.2"
flake8-black = "^0.1.1"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"
mypy = "^0.770"
nox = "^2019.11.9"
pre-commit = "^2.3.0"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
safety = "^1.9.0"
pytest-mock = "^3.3.1"

[tool.poetry.scripts]
convert = "csv2docx.cli:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["csv2docx"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[pytest]
mock_use_standalone_module = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
